+++ test/supercluster_immutable_test.dart
@@ -1,14 +1,17 @@
+import 'package:compute/compute.dart';
 import 'package:supercluster/supercluster.dart';
 import 'package:test/test.dart';
 
 import 'fixtures/fixtures.dart';
+import 'test_point.dart';
 
 void main() {
   SuperclusterImmutable<Map<String, dynamic>> supercluster(
-          List<Map<String, dynamic>> points,
-          {int? radius,
-          int? extent,
-          int? maxZoom}) =>
+    List<Map<String, dynamic>> points, {
+    int? radius,
+    int? extent,
+    int? maxZoom,
+  }) =>
       SuperclusterImmutable<Map<String, dynamic>>(
         getX: (json) {
           return json['geometry']?['coordinates'][0].toDouble();
@@ -21,11 +24,48 @@ void main() {
         maxZoom: maxZoom,
       )..load(points);
 
+  SuperclusterImmutable<TestPoint> supercluster2(
+          SuperclusterArgs superclusterArgs) =>
+      SuperclusterImmutable<TestPoint>(
+        getX: TestPoint.getX,
+        getY: TestPoint.getY,
+        minPoints: superclusterArgs.minPoints,
+        radius: superclusterArgs.radius,
+        extent: superclusterArgs.extent,
+        maxZoom: superclusterArgs.maxZoom,
+      )..load(superclusterArgs.points);
+
+  List<ImmutableLayerElement<T>> layerElementsAtZoom<T>(
+          SuperclusterImmutable<T> index, int zoom) =>
+      index.search(-180, -90, 180, 90, zoom).toList();
+
   test('removes existing points when calling load()', () {
     final index = supercluster(Fixtures.features);
     index.load([]);
-    expect(index.points, isEmpty);
+    expect(index.length, 0);
     expect(index.getLeaves(), isEmpty);
+  });
+
+  test('replacePoints', () async {
+    final testPoints = [
+      TestPoint(longitude: 9.203368, latitude: 45.460982), // Milano
+      TestPoint(longitude: 9.218777, latitude: 45.466276), // Milano east
+      TestPoint(longitude: 9.507878, latitude: 45.303647), // Lodi
+      TestPoint(longitude: 10.222456, latitude: 45.534990), // Brescia
+      TestPoint(longitude: 10.419535, latitude: 45.511298), // Bedizzole
+    ];
+    final index =
+        await compute(supercluster2, SuperclusterArgs(points: testPoints));
+    expect(index.getLeaves().any(testPoints.contains), isFalse);
+    index.replacePoints(testPoints);
+    expect(index.getLeaves().any((e) => !testPoints.contains(e)), isFalse);
+
+    for (int i = index.minZoom; i <= index.maxZoom + 1; i++) {
+      final pointsAtZoom = layerElementsAtZoom(index, i)
+          .whereType<ImmutableLayerPoint<TestPoint>>();
+      expect(pointsAtZoom.any((e) => !testPoints.contains(e.originalPoint)),
+          isFalse);
+    }
   });
 
   test('returns children of a cluster', () {
@@ -185,5 +225,23 @@ void main() {
     feature['geometry']['coordinates'][0] =
         feature['geometry']['coordinates'][0] + 1;
     expect(index.contains(feature), isFalse);
+  });
+}
+
+class SuperclusterArgs {
+  final List<TestPoint> points;
+  final int? minPoints;
+  final int? maxEntries;
+  final int? radius;
+  final int? extent;
+  final int? maxZoom;
+
+  SuperclusterArgs({
+    required this.points,
+    this.minPoints,
+    this.maxEntries,
+    this.radius,
+    this.extent,
+    this.maxZoom,
   });
 }